refactor(home): extract preview data, shared error card, and data model hook
- Extract preview cost data factory from useCostFilters to previewCostData.ts (~350 lines)
- Extract preview token data from HomeTokenCostPanel to previewTokenData.ts (~200 lines)
- Extract useHomeTokenCostDataModel hook from HomeTokenCostPanel (query orchestration)
- Create shared QueryErrorCard component, replace duplicated error cards in
  HomeTokenCostPanel and UsagePage
- Fix homeOverviewLayout subscribe to also listen for cross-tab storage events
- Change UsageSummary.cost_covered_success from optional to required (matches Rust backend)
- Expand TokenBreakdown tests: full mode, cache line visibility, zero/null edge cases
