+++ src/components/usage/__tests__/UsageLeaderboardTable.test.tsx
@@ -8,6 +8,7 @@ const SUMMARY: UsageSummary = {
   requests_with_usage: 0,
   requests_success: 0,
   requests_failed: 0,
+  cost_covered_success: 0,
   avg_duration_ms: null,
   avg_ttfb_ms: null,
   avg_output_tokens_per_second: null,
