+++ src/services/home/homeOverviewLayout.ts
@@ -9,6 +9,12 @@ function emit() {
   for (const listener of listeners) listener();
 }
 
+function handleStorageEvent(event: StorageEvent) {
+  if (event.key === HOME_OVERVIEW_LOGS_PRIMARY_LAYOUT_STORAGE_KEY) {
+    emit();
+  }
+}
+
 export function readHomeOverviewLogsPrimaryLayoutFromStorage(): boolean {
   if (typeof window === "undefined") return false;
 
@@ -35,6 +41,14 @@ export function writeHomeOverviewLogsPrimaryLayoutToStorage(enabled: boolean) {
 }
 
 export function subscribeHomeOverviewLogsPrimaryLayout(listener: Listener) {
+  if (listeners.size === 0 && typeof window !== "undefined") {
+    window.addEventListener("storage", handleStorageEvent);
+  }
   listeners.add(listener);
-  return () => listeners.delete(listener);
+  return () => {
+    listeners.delete(listener);
+    if (listeners.size === 0 && typeof window !== "undefined") {
+      window.removeEventListener("storage", handleStorageEvent);
+    }
+  };
 }
