+++ src/test/msw/state.ts
@@ -108,6 +108,7 @@ const DEFAULT_USAGE_SUMMARY: UsageSummary = {
   requests_with_usage: 0,
   requests_success: 0,
   requests_failed: 0,
+  cost_covered_success: 0,
   avg_duration_ms: null,
   avg_ttfb_ms: null,
   avg_output_tokens_per_second: null,
